輝度値変換用のテクスチャサイズをcompression値に応じて圧縮するようにした

+++ Source/UI_Blooman/Private/FakeBloomUI_Painter.cpp
@@ -12,7 +12,7 @@ void UFakeBloomUI_Painter::SetParameters(const FFakeBloomUI_BuildParameter* InBu
 
 void UFakeBloomUI_Painter::DrawImageToCenter(FPaintContext& Context,
                                              UObject* Image,
-                                             float Overhang,
+                                             int32 Overhang,
                                              const FVector2D& Scale,
                                              const FLinearColor& TintColor) const
 {
